【front】メディア作成・管理ページをmanage配下に統合しディレクトリ整理
