+++ frontend/src/components/templates/manage/Manage.module.scss
