+++ frontend/src/components/templates/manage/picture/create.tsx
@@ -1,8 +1,8 @@
 import { useState, ChangeEvent } from 'react'
 import { Channel } from 'types/internal/channel'
-import { MusicIn } from 'types/internal/media'
+import { PictureIn } from 'types/internal/media'
 import { Option } from 'types/internal/other'
-import { postMusicCreate } from 'api/internal/media/create'
+import { postPictureCreate } from 'api/internal/manage/create'
 import { FetchError } from 'utils/constants/enum'
 import { useIsLoading } from 'components/hooks/useIsLoading'
 import { useRequired } from 'components/hooks/useRequired'
@@ -10,7 +10,6 @@ import { useToast } from 'components/hooks/useToast'
 import Main from 'components/layout/Main'
 import Button from 'components/parts/Button'
 import Input from 'components/parts/Input'
-import CheckBox from 'components/parts/Input/CheckBox'
 import InputFile from 'components/parts/Input/File'
 import SelectBox from 'components/parts/Input/SelectBox'
 import Textarea from 'components/parts/Input/Textarea'
@@ -21,7 +20,7 @@ interface Props {
   channels: Channel[]
 }
 
-export default function MusicCreate(props: Props): React.JSX.Element {
+export default function PictureCreate(props: Props): React.JSX.Element {
   const { channels } = props
 
   const channelUlid = channels.find((c) => c.isDefault)?.ulid ?? ''
@@ -30,42 +29,37 @@ export default function MusicCreate(props: Props): React.JSX.Element {
   const { isLoading, handleLoading } = useIsLoading()
   const { isRequired, isRequiredCheck } = useRequired()
   const { toast, handleToast } = useToast()
-  const [values, setValues] = useState<MusicIn>({ channelUlid, publish: true, title: '', content: '', lyric: '', download: true })
+  const [values, setValues] = useState<PictureIn>({ channelUlid, publish: true, title: '', content: '' })
 
   const handlePublish = () => setValues({ ...values, publish: !values.publish })
   const handleSelect = (e: ChangeEvent<HTMLSelectElement>) => setValues({ ...values, [e.target.name]: e.target.value })
   const handleInput = (e: ChangeEvent<HTMLInputElement>) => setValues({ ...values, [e.target.name]: e.target.value })
   const handleText = (e: ChangeEvent<HTMLTextAreaElement>) => setValues({ ...values, [e.target.name]: e.target.value })
-  const handleCheck = (e: ChangeEvent<HTMLInputElement>) => setValues({ ...values, [e.target.name]: e.target.checked })
-  const handleFile = (files: File | File[]) => Array.isArray(files) || setValues({ ...values, music: files })
+  const handleFile = (files: File | File[]) => Array.isArray(files) || setValues({ ...values, image: files })
 
   const handleForm = async () => {
-    const { channelUlid, title, content, music } = values
-    if (!isRequiredCheck({ channelUlid, title, content, music })) return
+    const { channelUlid, title, content, image } = values
+    if (!isRequiredCheck({ channelUlid, title, content, image })) return
     handleLoading(true)
-    const ret = await postMusicCreate(values)
+    const ret = await postPictureCreate(values)
     handleLoading(false)
     if (ret.isErr()) {
       handleToast(FetchError.Post, true)
       return
     }
-    setValues({ channelUlid, publish: true, title: '', content: '', lyric: '', download: true })
+    setValues({ channelUlid, publish: true, title: '', content: '' })
     handleToast('作成しました', false)
   }
 
   return (
-    <Main title="Music" type="table" toast={toast} isFooter={false} button={<Button color="green" size="s" name="作成する" loading={isLoading} onClick={handleForm} />}>
-      <form method="POST" action="" encType="multipart/form-data">
+    <Main title="Picture" type="table" toast={toast} isFooter={false} button={<Button color="green" size="s" name="作成する" loading={isLoading} onClick={handleForm} />}>
+      <form method="POST" action="">
         <VStack gap="8">
           <ToggleCard label="公開する" isActive={values.publish} onClick={handlePublish} />
           <SelectBox label="チャンネル" name="channelUlid" value={values.channelUlid} options={channelOptions} onChange={handleSelect} />
           <Input label="タイトル" name="title" required={isRequired} onChange={handleInput} />
           <Textarea label="内容" name="content" required={isRequired} onChange={handleText} />
-          <Textarea label="歌詞" name="lyric" required={isRequired} onChange={handleText} />
-          <VStack gap="2">
-            <InputFile label="音楽" accept="audio/*" required={isRequired} onChange={handleFile} />
-            <CheckBox label="ダウンロード許可" name="download" defaultChecked onChange={handleCheck} />
-          </VStack>
+          <InputFile label="画像" accept="image/*" required={isRequired} onChange={handleFile} />
         </VStack>
       </form>
     </Main>
